added the nav for the chat menu
