+++ src/components/Chat/Chat.css
@@ -12,8 +12,7 @@
   display: flex;
   border-radius: 10px;
   height: 500px;
-  max-width: 600px;
-  width: 100%;
+  width: fit-content;
   margin: 0 10px;
   border: 2px solid rgb(246, 247, 248);
   overflow: hidden;
@@ -23,11 +22,15 @@
   background-color: white;
 }
 
+.chat-right-side {
+  width: 300px;
+}
+
 /* CHAT ASIDE =========================== */
 
 .chat-aside {
   background-color: white;
-  width: 250px !important;
+  width: 200px;
   padding: 20px;
   border: 1px solid rgb(246, 247, 248);
 }
@@ -55,6 +58,13 @@
   border-bottom: 1px solid black;
   margin-bottom: 5px;
   cursor: pointer;
+  display: flex;
+  height: 30px;
+  align-items: center;
+  overflow-x: scroll;
+}
+.chat-aside-room > p {
+  margin-right: 5px;
 }
 .chat-aside-room:hover {
   background-color: rgb(246, 247, 248);
@@ -65,13 +75,13 @@
 .chat-main {
   background-color: rgb(246, 247, 248);
   width: 100%;
-  padding: 20px;
 }
 
 .chat-main-nav {
   display: flex;
   justify-content: space-between;
   align-items: center;
+  padding: 10px;
 }
 
 .chat-main-icons > * {
@@ -192,18 +202,20 @@
 .messages-container {
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
+  height: 90%;
   width: 100%;
   padding: 10px 0;
+  margin-bottom: 10px;
 }
 
 .messages-form {
   width: 100%;
+  justify-self: flex-end;
   display: flex;
 }
 .messages-form > input {
   width: 100%;
+  padding: 10px;
 }
 
 .messages-text-container {
@@ -218,7 +230,7 @@
   background-color: rgb(255, 167, 152);
   border: 1px solid black;
   border-radius: 20px;
-  padding: 5px 10px;
+  padding: 5px 15px;
   width: fit-content;
   margin: 5px;
 }
